% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dkt.R
\name{dkt}
\alias{dkt}
\title{Desikan-Killiany-Tourville Cortical Atlas}
\usage{
dkt()
}
\value{
A \link[ggseg.formats:ggseg_atlas]{ggseg.formats::ggseg_atlas} object (cortical).
}
\description{
Cortical parcellation with 32 regions per hemisphere based on
the Desikan-Killiany-Tourville labeling protocol. Contains both
2D polygon geometry for \code{\link[ggseg:ggbrain]{ggseg::geom_brain()}} and 3D vertex
indices for \code{\link[ggseg3d:ggseg3d]{ggseg3d::ggseg3d()}}.
}
\examples{
dkt()
plot(dkt())
}
\references{
Klein A, Tourville J (2012). 101 labeled brain
images and a consistent human cortical labeling protocol.
\emph{Frontiers in Neuroscience}, 6:171.
\doi{10.3389/fnins.2012.00171}
}
\seealso{
Other ggseg_atlases: 
\code{\link{brainstem}()},
\code{\link{destrieux}()},
\code{\link{hcpa}()},
\code{\link{hippoamyg}()},
\code{\link{hypothalamus}()},
\code{\link{thalamus}()}

Other cortical_atlases: 
\code{\link{destrieux}()}

Other freesurfer_atlases: 
\code{\link{brainstem}()},
\code{\link{destrieux}()},
\code{\link{hcpa}()},
\code{\link{hippoamyg}()},
\code{\link{hypothalamus}()},
\code{\link{thalamus}()}
}
\concept{cortical_atlases}
\concept{freesurfer_atlases}
\concept{ggseg_atlases}
