% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{destrieux}
\alias{destrieux}
\title{Destrieux Cortical Atlas}
\usage{
destrieux()
}
\value{
A \link[ggseg.formats:ggseg_atlas]{ggseg.formats::ggseg_atlas} object (cortical).
}
\description{
Brain atlas for the Destrieux cortical parcellation
(aparc.a2009s) with 75 regions per hemisphere. Contains 2D
polygon geometry for \code{\link[ggseg:ggbrain]{ggseg::geom_brain()}}.
}
\examples{
destrieux()
plot(destrieux())
}
\references{
Destrieux C, Fischl B, Dale A, Halgren E (2010).
Automatic parcellation of human cortical gyri and sulci
using standard anatomical nomenclature. \emph{NeuroImage},
53(1), 1-15. \doi{10.1016/j.neuroimage.2010.06.010}
}
\seealso{
Other ggseg_atlases: 
\code{\link{dkt}()},
\code{\link{hcpa}()}

Other cortical_atlases: 
\code{\link{dkt}()}

Other freesurfer_atlases: 
\code{\link{dkt}()},
\code{\link{hcpa}()}
}
\concept{cortical_atlases}
\concept{freesurfer_atlases}
\concept{ggseg_atlases}
