% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggsegFreeSurfer-package.R
\docType{package}
\name{ggsegFreeSurfer-package}
\alias{ggsegFreeSurfer}
\alias{ggsegFreeSurfer-package}
\title{ggsegFreeSurfer: FreeSurfer Brain Atlases for the 'ggseg' Ecosystem}
\description{
FreeSurfer cortical and subcortical brain atlases for the 'ggseg' ecosystem. Includes the Desikan-Killiany-Tourville cortical atlas (Klein & Tourville 2012), the Destrieux cortical atlas (Destrieux et al. 2010), and the HCP subcortical atlas with anterior/posterior hippocampus subdivisions. Provides unified 'ggseg_atlas' objects with 2D polygon geometry and 3D mesh or vertex data, for use with 'ggseg' and 'ggseg3d'.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/ggsegverse/ggsegFreeSurfer}
  \item Report bugs at \url{https://github.com/ggsegverse/ggsegFreeSurfer/issues}
}

}
\author{
\strong{Maintainer}: Athanasia Mo Mowinckel \email{a.m.mowinckel@psykologi.uio.no} (\href{https://orcid.org/0000-0002-5756-0223}{ORCID})

Authors:
\itemize{
  \item Didac Vidal-Piñeiro \email{d.v.pineiro@psykologi.uio.no} (\href{https://orcid.org/0000-0001-9997-9156}{ORCID})
}

}
\keyword{internal}
