% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{hcpa}
\alias{hcpa}
\title{HCP Subcortical Atlas}
\usage{
hcpa()
}
\value{
A \link[ggseg.formats:ggseg_atlas]{ggseg.formats::ggseg_atlas} object (subcortical).
}
\description{
Human Connectome Project subcortical atlas with aseg regions
and anterior/posterior hippocampus subdivisions. Contains 3D
mesh geometry for \code{\link[ggseg3d:ggseg3d]{ggseg3d::ggseg3d()}}.
}
\examples{
hcpa()
}
\seealso{
Other ggseg_atlases: 
\code{\link{destrieux}()},
\code{\link{dkt}()}

Other freesurfer_atlases: 
\code{\link{destrieux}()},
\code{\link{dkt}()}
}
\concept{freesurfer_atlases}
\concept{ggseg_atlases}
\concept{subcortical_atlases}
